+++ EliteJournalReader/Events/ScanEvent.cs
@@ -54,11 +54,11 @@ namespace EliteJournalReader.Events
                 base.Initialize(evt);
                 BodyName = evt.Value<string>("BodyName");
                 StarType = evt.Value<string>("StarType");
-                StellarMass = evt.Value<decimal?>("StellarMass");
-                Radius = evt.Value<decimal?>("StellarMass");
-                AbsoluteMagnitude = evt.Value<decimal?>("StellarMass");
-                OrbitalPeriod = evt.Value<decimal>("StellarMass");
-                RotationPeriod = evt.Value<decimal>("StellarMass");
+                StellarMass = evt.Value<double?>("StellarMass");
+                Radius = evt.Value<double?>("StellarMass");
+                AbsoluteMagnitude = evt.Value<double?>("StellarMass");
+                OrbitalPeriod = evt.Value<double>("StellarMass");
+                RotationPeriod = evt.Value<double>("StellarMass");
                 Rings = evt["Rings"]?.ToObject<PlanetRing[]>();
 
                 TidalLock = evt.Value<bool?>("TidalLock") ?? false;
@@ -66,22 +66,22 @@ namespace EliteJournalReader.Events
                 PlanetClass = evt.Value<string>("PlanetClass");
                 Atmosphere = evt.Value<string>("Atmosphere");
                 Volcanism = evt.Value<string>("Volcanism");
-                MassEM = evt.Value<decimal?>("MassEM");
-                SurfaceGravity = evt.Value<decimal?>("SurfaceGravity");
-                SurfaceTemperature = evt.Value<decimal?>("SurfaceTemperature");
-                SurfacePressure = evt.Value<decimal?>("SurfacePressure");
+                MassEM = evt.Value<double?>("MassEM");
+                SurfaceGravity = evt.Value<double?>("SurfaceGravity");
+                SurfaceTemperature = evt.Value<double?>("SurfaceTemperature");
+                SurfacePressure = evt.Value<double?>("SurfacePressure");
                 Landable = evt.Value<bool?>("Landable") ?? false;
-                Materials = evt["Materials"]?.ToObject<Dictionary<string, decimal>>();
+                Materials = evt["Materials"]?.ToObject<Dictionary<string, double>>();
             }
 
             public string BodyName { get; set; }
-            public decimal DistanceFromArrivalLS { get; set; }
+            public double DistanceFromArrivalLS { get; set; }
             public string StarType { get; set; }
-            public decimal? StellarMass { get; set; }
-            public decimal? Radius { get; set; }
-            public decimal? AbsoluteMagnitude { get; set; }
-            public decimal OrbitalPeriod { get; set; }
-            public decimal RotationPeriod { get; set; }
+            public double? StellarMass { get; set; }
+            public double? Radius { get; set; }
+            public double? AbsoluteMagnitude { get; set; }
+            public double OrbitalPeriod { get; set; }
+            public double RotationPeriod { get; set; }
             public PlanetRing[] Rings { get; set; }
 
             public bool? TidalLock { get; set; }
@@ -89,12 +89,12 @@ namespace EliteJournalReader.Events
             public string PlanetClass { get; set; }
             public string Atmosphere { get; set; }
             public string Volcanism { get; set; }
-            public decimal? MassEM { get; set; } // not in description of event
-            public decimal? SurfaceGravity { get; set; } // not in description of event
-            public decimal? SurfaceTemperature { get; set; }
-            public decimal? SurfacePressure { get; set; }
+            public double? MassEM { get; set; } // not in description of event
+            public double? SurfaceGravity { get; set; } // not in description of event
+            public double? SurfaceTemperature { get; set; }
+            public double? SurfacePressure { get; set; }
             public bool? Landable { get; set; }
-            public Dictionary<string, decimal> Materials { get; set; }
+            public Dictionary<string, double> Materials { get; set; }
         }
     }
 
@@ -102,8 +102,8 @@ namespace EliteJournalReader.Events
     {
         public string Name;
         public string RingClass;
-        public decimal MassMT;
-        public decimal InnerRad;
-        public decimal OuterRad;
+        public double MassMT;
+        public double InnerRad;
+        public double OuterRad;
     }
 }
